lib: Add littlefs.
This adds the littlefs library. This is used to provide the file
system on the SPIKE Prime hub, the SPIKE Essential hub,
and the MINDSTORMS Robot Inventor hub.

We will use littlefs v1 for the moment, in order to be compatible
with the original firmware. This way, the dual-boot version can use
the existing filesystem without having to format it.